integrate options-server into test-options.py
